fix sigma

+++ moses/src/SpanLengthEstimator.cpp
@@ -70,7 +70,7 @@ public:
   virtual void FinishedAdds(unsigned ruleCount) {
     //m_sigma = max(1.0f, sqrt(m_averageSquare - m_average*m_average));
     CHECK(ruleCount > 0);
-    m_sigma=max(1.0f,sqrt(m_averageSquare - m_average*m_average+m_ISIk/(1+m_ISIk/ruleCount)*m_averageSquare));
+    m_sigma=max(1.0f,sqrt(m_averageSquare - m_average*m_average+m_ISIk/(1+(float)m_ISIk/ruleCount)*m_averageSquare));
     m_logSigma = log(m_sigma);
   }
 };
